fixed broken test setup; detoxed
