+++ tests/test_composite_corupture_sections.py
@@ -1,6 +1,6 @@
 import json
 from unittest.mock import patch
-import pathlib
+
 import pytest
 from graphene.test import Client
 
@@ -11,12 +11,12 @@ query {
   filter_rupture_sections(
       filter:{
       model_id: "NSHM_v1.0.4",
-      corupture_fault_names: ["Alpine: George landward"],
+      corupture_fault_names: ["Masterton"],
       location_ids: [],
       fault_system: "CRU",
       radius_km: 100
-      minimum_rate: 1.0e-20
-      minimum_mag: 6.2
+      minimum_rate: 1.0e-9
+      minimum_mag: 7.2
       filter_set_options: {
         multiple_locations:INTERSECTION
         multiple_faults: INTERSECTION
@@ -38,26 +38,18 @@ query {
 }
 """
 
+
 @pytest.fixture(scope='class')
 def client():
     return Client(schema_root)
 
 
-@pytest.fixture(autouse=True)
-def archive_fixture(monkeypatch):
-    # archive_path = pathlib.Path(__name__).parent.parent / "data_store/test/fixtures" / "TinyCompositeSolution.zip"
-    archive_path = pathlib.Path(__name__).parent.parent / "WORKING" / "NSHM_v1.0.4_CompositeSolution.zip"
-    assert archive_path.exists()
-    monkeypatch.setenv("COMPOSITE_ARCHIVE_PATH", str(archive_path))
-    yield archive_path
-
 @patch(
     'solvis_graphql_api.composite_solution.cached.get_fault_name_rupture_ids',
     lambda *args, **kwargs: [n for n in range(350, 390)],
 )
-@patch('solvis_graphql_api.composite_solution.cached.RESOLVE_LOCATIONS_INTERNALLY', False)
+@patch('solvis_graphql_api.composite_solution.cached.RESOLVE_LOCATIONS_INTERNALLY', True)
 class TestFilterRptureSections:
-
     def test_get_fault_surfaces_styled(self, client, archive_fixture):
 
         executed = client.execute(
@@ -118,13 +110,13 @@ class TestFilterRptureSections:
         f1 = json.loads(executed['data']['filter_rupture_sections']['fault_surfaces'])
         assert f1['features'][0] == f0
 
-    def test_get_min_magnitude(self, client):
+    def test_get_min_magnitude(self, client, archive_fixture):
 
         executed = client.execute(QUERY)
         print(executed)
         assert pytest.approx(executed['data']['filter_rupture_sections']['min_magnitude']) == 7.240527629852295
 
-    def test_get_mfd_histogram(self, client):
+    def test_get_mfd_histogram(self, client, archive_fixture):
         executed = client.execute(
             QUERY.replace("# MFD", "mfd_histogram{ bin_center rate cumulative_rate}"),
         )
