+++ tests/test_rupture_sorting.py
@@ -160,7 +160,7 @@ def verify_sorted_edges(edges: List[Dict], fields: List[SortedField]):
     ],
 )
 @patch('solvis_graphql_api.composite_solution.cached.RESOLVE_LOCATIONS_INTERNALLY', True)
-def test_sorting_and_binning_magnitude(client, query, variable_values, sort_expr, expected):
+def test_sorting_and_binning_magnitude(client, query, variable_values, sort_expr, expected, archive_fixture):
     print(query)
     executed = client.execute(
         query.replace("# SORT_BY", sort_expr).replace("#FIRST", "30"),
@@ -186,7 +186,7 @@ def test_sorting_and_binning_magnitude(client, query, variable_values, sort_expr
     ],
 )
 @pytest.mark.skip("use this for test hacking")
-def test_magnitude_binning(client, query, variable_values, sort_expr, expected):
+def test_magnitude_binning(client, query, variable_values, sort_expr, expected, archive_fixture):
     print(query)
 
     LIMIT = 200
